+++ supabase/migrations/20251130110000_processing_status.sql
@@ -20,3 +20,7 @@ COMMENT ON COLUMN items.processing_status IS 'Background processing status: pend
 
 
 
+
+
+
+